+++ src/main/java/org/junit/rules/ExternalResource.java
@@ -1,0 +1,11 @@
+package org.junit.rules;
+
+/**
+ * Fake ExternalResource class from junit4 to get rid of the direct
+ * junit4 dependency of MockWebServer.
+ *
+ * {@see https://github.com/square/okhttp/issues/4667}
+ * {@see https://github.com/square/okhttp/issues/2635}
+ */
+public class ExternalResource {
+}